17-12-11-b

+++ concurrency_tools/concurrency_tools.h
@@ -172,7 +172,7 @@ template<typename T, uint32_t block, uint32_t index_sz>
 class fixed_sz_deque {
 private:
 	std::atomic<T*> index_array[index_sz] = { nullptr };
-	std::atomic<uint32_t> first_free = 0;
+	std::atomic<uint64_t> first_free = 0;
 	std::atomic<uint32_t> first_free_index = 1;
 
 	void create_new_block();
